Remove env file

+++ BackEnd/controllers/product.controller.js
@@ -33,12 +33,17 @@ export const createProducts =  async (req,res) => {
 
 export const deleteProduct = async (req,res) => {
     const {id} = req.params;
+
+    if(!mongoose.Types.ObjectId.isValid(id)){
+        return res.status(404).json({success:false, message: "Invalid Product id"});
+    }
+    
     try {
         await Product.findByIdAndDelete(id);
         res.status(200).json({success:true, message:"Product Delted"});
     } catch (error) {
         console.error("Error in deleting product:", error.message);
-        res.status(400).json({success:false, message:"Product not Found"});
+        res.status(500).json({success:false, message:"Server Error"});
     }
     console.log(`Product with ${id} has been deleted:`);
